// src/types/types.ts
export type PostType = {
    id: number,
    message: string,
    likeCount: number
}
export type ContactType = {
    github: string
    vk: string
    facebook: string
    instagram: string
    twitter: string
    website: string
    youtube: string
    mainLink: string
}
export type PhotosType = {
    small: string
    large: string
}
export type ProfileType = {
    userId: number
    lookingForAJob: boolean
    lookingForAJobDescription: string
    fullName: string
    contacts: ContactType
    photos: PhotosType
}


export type UsersType = {
    id: number
    name: string
    status: string
    photos: PhotosType
    followed: boolean
}